Added Forest
Added the forest where user is meant to go get the sword to fight baddies
Added a spawn system so upon player's return to the city, the last entered door is where player appears.
Lowered view distance in forest to add to mystery/maze feel + help performance.

Immediate TODO: make the "baddoor" area that will be the place of battle